+++ djangae/db/migrations/mapper_library.py
@@ -7,16 +7,15 @@
 
 import copy
 import cPickle
+import itertools
 import logging
 from datetime import datetime
 
-from google.appengine.api import datastore_errors
-from google.appengine.api.taskqueue.taskqueue import _DEFAULT_QUEUE
-
 from djangae.db.backends.appengine import rpc
 from django.conf import settings
-from django.utils import six
 from django.utils.six.moves import range
+from google.appengine.api import datastore_errors
+from google.appengine.api.taskqueue.taskqueue import _DEFAULT_QUEUE
 from google.appengine.ext import deferred
 from google.appengine.runtime import DeadlineExceededError
 
@@ -26,171 +25,62 @@ class Redefer(Exception):
     pass
 
 
-def _mid_string(string1, string2):
-    """ Given 2 unicode strings, return the string that is alphabetically half way between them. """
-    # Put the strings in order, so the lowest one is lhs
-    lhs = min(string1, string2)
-    rhs = max(string1, string2)
-    # Pad out the shorter string so that they're both the same length
-    longest_length = max(len(lhs), len(rhs))
-    lhs = lhs.ljust(longest_length, "\0")
-    # For each position in the strings, find the mid character
-    mid = []
-    for l, r in zip(lhs, rhs):
-        l = ord(l)
-        r = ord(r)
-        mid.append(l + (r - l) / 2)
-    # Note that some of the numbers might be invalid unicode values, but for the purposes of
-    # filtering Datastore keys it doesn't matter
-    result = u"".join([unichr(x) for x in mid])
-    # Strings starting with a double underscore are not valid Datastore keys
-    if result.startswith(u"__"):
-        result = u"_`" + result[2:]
-    return result
+
+def pairwise(iterable):
+    "s -> (s0,s1), (s1,s2), (s2, s3), ..."
+    # Copied from Python docs.
+    a, b = itertools.tee(iterable)
+    next(b, None)
+
+    return itertools.izip(a, b)
 
 
-def _next_string(string):
-    """ Given a string (or unicode), return the alphabetically next string. """
-    # Note that in python 2 at least, unicode is 16 bit, and therefore some characters (e.g. emoji)
-    # are encoded as 2 characters, so when we slice the last "character" off the string we're
-    # actually getting half a character, and then when we increment it we're possibly creating an
-    # invalid character, but for the purpose of ordering Datastore keys it shouldn't matter
-    try:
-        # Try to increment the last character by 1
-        return string[:-1] + unichr(ord(string[-1]) + 1)
-    except ValueError:
-        # If the last character was already the highest possible unicode value, then instead add
-        # another character to the string
-        return string + unichr(1)
+def good_fit_step(x, y):
+    """Returns a step interval to use when selecting keys.
 
-
-def _next_key(key):
+    The step will ensure that we create no more than num_shards.
     """
-        Given a key, this returns key + 1. In the case of key names
-        we simply calculate the next alphabetical key
+    step, remainder = divmod(x, y)
+
+    if remainder:
+        step = (x + remainder) // y
+
+    if step == 0:
+        step = 1
+
+    return step
+
+
+def generate_shards(keys, shard_count):
+    """Returns a list of key pairs, where the last pair is like (<key>, None).
+
+    Each (start, stop) pair can be used to filter entities like:
+
+      WHERE __key__ >= start AND __key__ < stop
+
+    (Except for the last pair, when stop is None). Note this will never return
+    more than shard_count pairs.
+
+    >>> keys = [1, 2, 3, 4, 5]
+    >>> generate_shards(keys, shard_count=1)
+    [(1, None)]
+    >>> generate_shards(keys, shard_count=2)
+    [(1, 4), (4, None)]
+    >>> generate_shards(keys, shard_count=5)
+    [(1, 2), (2, 3), (3, 4), (4, 5), (5, None)]
     """
-    val = key.id_or_name()
-    if isinstance(val, six.string_types):
-        return rpc.Key.from_path(
-            key.kind(),
-            _next_string(val),
-            namespace=key.namespace()
-        )
-    else:
-        return rpc.Key.from_path(
-            key.kind(),
-            val + 1,
-            namespace=key.namespace()
-        )
+    step = good_fit_step(len(keys), shard_count)
 
-
-def _mid_key(key1, key2):
-    """
-        Given two keys, this function returns the key mid-way between them
-        - this needs some thought on how to do this with strings
-    """
-    key1_val = key1.id_or_name()
-    key2_val = key2.id_or_name()
-
-    if type(key1_val) != type(key2_val):
-        raise NotImplementedError(
-            "Sharding of entities with mixed integer and string types is not yet supported."
-        )
-
-    if isinstance(key1_val, six.string_types):
-        mid_id_or_name = _mid_string(key1_val, key2_val)
-    else:
-        mid_id_or_name = key1_val + ((key2_val - key1_val) // 2)
-
-    return rpc.Key.from_path(
-        key1.kind(),
-        mid_id_or_name,
-        namespace=key1.namespace()
-    )
-
-
-def _get_range(key1, key2):
-    """ Given 2 Datastore keys, return the range that their IDs span.
-        E.g. if the IDs are 7 and 100, then the range is 93.
-        Works for string-based keys as well, but returns a string representation of the difference.
-    """
-    val1 = key1.id_or_name()
-    val2 = key2.id_or_name()
-    if type(val1) != type(val2):
-        raise Exception("Cannot calculate range between keys of different types.")
-
-    # Otherwise, the values are strings...
-    # Put the strings in order, so the lowest one is lhs
-    lhs = min(val1, val2)
-    rhs = max(val1, val2)
-
-    if isinstance(lhs, (int, long)):
-        return rhs - lhs
-
-    # Pad out the shorter string so that they're both the same length
-    longest_length = max(len(lhs), len(rhs))
-    lhs = lhs.ljust(longest_length, "\0")
-
-    max_unicode = 0x10FFFF
-
-    # This is based on the positional numeral system solution described here:
-    # https://stackoverflow.com/a/41492405/48362
-    # But adapted to deal with unicode characters.
-
-    lhs_value = sum((max_unicode ** n) * ord(c) for n, c in enumerate(reversed(lhs)))
-    rhs_value = sum((max_unicode ** n) * ord(c) for n, c in enumerate(reversed(rhs)))
-
-    return rhs_value - lhs_value
-
-
-def _generate_shards(keys, shard_count):
-    """
-        Given a set of keys with:
-        - The first key being the lowest in the range
-        - The last key being the highest in the range
-        - The other keys being evenly distributed (e.g. __scatter__)
-
-        This function returns a list of [start_key, end_key] shards to cover the range
-
-        This may not return shard_count shards if there aren't enough split points
-    """
     keys = sorted(keys)  # Ensure the keys are sorted
+    keys = keys[::step]
 
-    # Special case single key
-    if shard_count == 1:
-        return [[keys[0], keys[-1]]]
-    elif shard_count < len(keys):
-        # If there are more split point keys than we need then We have to calculate a
-        # stride to skip some of the split point keys to return shard_count shards
-        index_stride = len(keys) / float(shard_count)
-        keys = [keys[int(round(index_stride * i))] for i in range(1, shard_count)]
+    # The last key pair looks like (<key>, None). See where we build the entity
+    # query in ShardedTaskManager.run_shard().
+    keys.append(None)
 
-    shards = []
-    for i in range(len(keys) - 1):
-        shards.append([keys[i], keys[i + 1]])
+    shards = list(pairwise(keys))
 
     return shards
-
-
-def _find_largest_shard(shards):
-    """
-        Given a list of shards, where each shard is a pair of (lowest_key, highest_key),
-        return the shard with the largest ID range
-    """
-    largest_shard = None
-    range_of_largest_shard = None
-
-    for shard in shards:
-        if largest_shard is None:
-            largest_shard = shard
-            range_of_largest_shard = _get_range(shard[0], shard[1])
-        else:
-            this_range = _get_range(shard[0], shard[1])
-            if this_range > range_of_largest_shard:
-                largest_shard = shard
-                range_of_largest_shard = _get_range(shard[0], shard[1])
-
-    return largest_shard
 
 
 def shard_query(query, shard_count):
@@ -223,53 +113,10 @@ def shard_query(query, shard_count):
         if keys[0] != min_id:
             keys.insert(0, min_id)
 
-        if keys[-1] != max_id or min_id == max_id:
+        if keys[-1] != max_id:
             keys.append(max_id)
 
-    # We generate as many shards as we can, but if it's not enough then we
-    # iterate, splitting the largest shard into two shards until either:
-    # - we hit the desired shard count
-    # - we can't subdivide anymore
-    shards = _generate_shards(keys, shard_count)
-    while True:
-        if len(shards) >= shard_count:
-            break
-
-        # If we don't have enough shards, divide the largest key range until we have enough
-        largest_shard = _find_largest_shard(shards)
-
-        # OK we can't shard anymore, just bail
-        if largest_shard[0] == largest_shard[1]:
-            break
-
-        left_shard = [
-            largest_shard[0],
-            _mid_key(largest_shard[0], largest_shard[1])
-        ]
-
-        right_shard = [
-            _next_key(_mid_key(largest_shard[0], largest_shard[1])),
-            largest_shard[1]
-        ]
-
-        # We already have these shards, just give up now
-        if left_shard in shards and right_shard in shards:
-            break
-
-        shards.remove(largest_shard)
-        if left_shard not in shards:
-            shards.append(left_shard)
-
-        if right_shard not in shards:
-            shards.append(right_shard)
-        shards.sort()
-
-    assert len(shards) <= shard_count
-
-    # We shift the end keys in each shard by one, so we can
-    # do a >= && < query
-    for shard in shards:
-        shard[1] = _next_key(shard[1])
+    shards = generate_shards(keys, shard_count)
 
     return shards
 
@@ -348,10 +195,17 @@ class ShardedTaskMarker(rpc.Entity):
 
         # Copy the query so that we can re-defer the original, unadulterated version, because once
         # we've applied limits and ordering to the query it causes pickle errors with defer.
+        start_key, stop_key = shard
         query = copy.deepcopy(original_query)
         query.Order("__key__")
-        query["__key__ >="] = shard[0]
-        query["__key__ <"] = shard[1]
+        query["__key__ >="] = start_key
+
+        # The last key pair looks like (<key>, None). When querying entities,
+        # the first key is inclusive and the second key is exclusive. But if
+        # the stop key is None then this must be the last shard so we omit the
+        # upper limit clause.
+        if stop_key is not None:
+            query["__key__ <"] = stop_key
 
         num_entities_processed = 0
         try: